moved support traits and classes to specific namespace
