+++ tests/Any/AnyOfTest.php
@@ -7,10 +7,10 @@ use Maxonfjvipon\ElegantElephant\Any\AnyOf;
 use Maxonfjvipon\ElegantElephant\Arr\ArrOf;
 use Maxonfjvipon\ElegantElephant\Logic\LogicOf;
 use Maxonfjvipon\ElegantElephant\Num\NumOf;
-use Maxonfjvipon\ElegantElephant\Tests\Support\IsArr;
-use Maxonfjvipon\ElegantElephant\Tests\Support\IsLogic;
-use Maxonfjvipon\ElegantElephant\Tests\Support\IsNum;
-use Maxonfjvipon\ElegantElephant\Tests\Support\IsTxt;
+use Maxonfjvipon\ElegantElephant\Support\IsArr;
+use Maxonfjvipon\ElegantElephant\Support\IsLogic;
+use Maxonfjvipon\ElegantElephant\Support\IsNum;
+use Maxonfjvipon\ElegantElephant\Support\IsTxt;
 use Maxonfjvipon\ElegantElephant\Tests\TestCase;
 use Maxonfjvipon\ElegantElephant\Txt\TxtOf;
 use PHPUnit\Framework\Constraint\Count;
