+++ tests/Num/NumCondTest.php
@@ -7,7 +7,7 @@ use Maxonfjvipon\ElegantElephant\Logic\LogicOf;
 use Maxonfjvipon\ElegantElephant\Num\NumCond;
 use Maxonfjvipon\ElegantElephant\Num\NumOf;
 use Maxonfjvipon\ElegantElephant\Num\SumOf;
-use Maxonfjvipon\ElegantElephant\Tests\Support\IsNum;
+use Maxonfjvipon\ElegantElephant\Support\IsNum;
 use Maxonfjvipon\ElegantElephant\Tests\TestCase;
 use PHPUnit\Framework\Constraint\IsEqual;
